returns the number of bits you would need to flip

+++ 0x14-bit_manipulation/5-flip_bits.c
@@ -9,7 +9,7 @@
 unsigned int flip_bits(unsigned long int n, unsigned long int m)
 {
 
-	unsigned int aux, cont;
+	unsigned long int aux, cont;
 
 	aux = n ^ m;
